+++ Kernel/Output/HTML/Dashboard/TicketGeneric.pm
@@ -1337,6 +1337,9 @@ sub Run {
             elsif ( $HeaderColumn eq 'CustomerUserID' ) {
                 $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Customer User ID');
             }
+            elsif ( $HeaderColumn eq 'AccountedTime' ) {
+                $TranslatedWord = $LayoutObject->{LanguageObject}->Translate('Accounted time');
+            }
             else {
                 $TranslatedWord = $LayoutObject->{LanguageObject}->Translate($HeaderColumn);
             }